ironaccord-bot/requirements.txt

# Core Discord Bot Libraries
discord.py==2.3.2
python-dotenv==1.0.1

# Asynchronous Database Connector
aiomysql==0.2.0

# Asynchronous HTTP Requests for Ollama API
httpx==0.27.0

# --- NEW: RAG and Vector Store Libraries ---
langchain==0.1.16
langchain-community==0.0.32
faiss-cpu==1.7.4 # For local vector storage
tqdm==4.66.4
markdown

# Testing Framework
pytest==8.2.2

# Vector store
chromadb
